Markov's modeling and presentation of expected values
Minor changes on the docstrings and comments. Implementation of the level linearity on the reference level range test. Making more selective the error launching when recognizing results. Implementation of the real-time presentation system of the results. Implementation of an automatic way of modeling Markov's chains for the variability of the result, and calculating the limit distribution and expected value.

Signed-off-by: jfBranch <jfmaldonado77v@gmail.com>